requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
pymysql==1.1.1
aiomysql==0.2.0
aiosqlite==0.20.0
greenlet==3.1.1
alembic==1.13.3

# Authentication
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# AWS S3
boto3==1.35.0

# Google AI (Gemini)
google-cloud-aiplatform==1.67.0
google-generativeai==0.8.0
vertexai==1.67.0

# HTTP Client
httpx==0.27.0
aiohttp==3.10.5

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0

# Logging
loguru==0.7.2

# CORS
starlette==0.38.0

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.0

# Code Quality
black==24.8.0
ruff==0.6.5
mypy==1.11.2
